+++ yrs/src/block_iter.rs
@@ -9,8 +9,8 @@ use crate::{Assoc, ID};
 #[derive(Debug, Clone)]
 pub(crate) struct BlockIter {
     branch: BranchPtr,
-    index: u32,
-    rel: u32,
+    index: u64,
+    rel: u64,
     next_item: Option<ItemPtr>,
     curr_move: Option<ItemPtr>,
     curr_move_start: Option<ItemPtr>,
@@ -37,7 +37,7 @@ impl BlockIter {
     }
 
     #[inline]
-    pub fn rel(&self) -> u32 {
+    pub fn rel(&self) -> u64 {
         self.rel
     }
 
@@ -69,7 +69,7 @@ impl BlockIter {
         }
     }
 
-    pub fn move_to(&mut self, index: u32, txn: &mut TransactionMut) {
+    pub fn move_to(&mut self, index: u64, txn: &mut TransactionMut) {
         if index > self.index {
             if !self.try_forward(txn, index - self.index) {
                 panic!("Block iter couldn't move forward");
@@ -79,7 +79,7 @@ impl BlockIter {
         }
     }
 
-    fn can_forward(&self, ptr: Option<ItemPtr>, len: u32) -> bool {
+    fn can_forward(&self, ptr: Option<ItemPtr>, len: u64) -> bool {
         if !self.reached_end || self.curr_move.is_some() {
             if len > 0 {
                 return true;
@@ -95,13 +95,13 @@ impl BlockIter {
         false
     }
 
-    pub fn forward<T: ReadTxn>(&mut self, txn: &T, len: u32) {
+    pub fn forward<T: ReadTxn>(&mut self, txn: &T, len: u64) {
         if !self.try_forward(txn, len) {
             panic!("Length exceeded")
         }
     }
 
-    pub fn try_forward<T: ReadTxn>(&mut self, txn: &T, mut len: u32) -> bool {
+    pub fn try_forward<T: ReadTxn>(&mut self, txn: &T, mut len: u64) -> bool {
         if len == 0 && self.next_item.is_none() {
             return true;
         }
@@ -182,7 +182,7 @@ impl BlockIter {
         }
     }
 
-    pub fn backward<T: ReadTxn>(&mut self, txn: &mut T, mut len: u32) {
+    pub fn backward<T: ReadTxn>(&mut self, txn: &mut T, mut len: u64) {
         if self.index < len {
             panic!("Length exceeded");
         }
@@ -296,7 +296,7 @@ impl BlockIter {
         self.reached_end = false;
     }
 
-    pub fn delete(&mut self, txn: &mut TransactionMut, mut len: u32) {
+    pub fn delete(&mut self, txn: &mut TransactionMut, mut len: u64) {
         let mut item = self.next_item;
         if self.index + len > self.branch.content_len() {
             panic!("Length exceeded");
@@ -357,15 +357,15 @@ impl BlockIter {
         self.next_item = item;
     }
 
-    pub(crate) fn slice<T: ReadTxn>(&mut self, txn: &T, buf: &mut [Value]) -> u32 {
-        let mut len = buf.len() as u32;
+    pub(crate) fn slice<T: ReadTxn>(&mut self, txn: &T, buf: &mut [Value]) -> u64 {
+        let mut len = buf.len() as u64;
         if self.index + len > self.branch.content_len() {
             return 0;
         }
         self.index += len;
         let mut next_item = self.next_item;
         let encoding = txn.store().options.offset_kind;
-        let mut read = 0u32;
+        let mut read = 0u64;
         while len > 0 {
             if !self.reached_end {
                 while let Some(item) = next_item {
@@ -379,7 +379,7 @@ impl BlockIter {
                             let r = item
                                 .content
                                 .read(self.rel as usize, &mut buf[read as usize..])
-                                as u32;
+                                as u64;
                             read += r;
                             len -= r;
                             if self.rel + r == item.content_len(encoding) {
